fix: integer overflow for large vram gpus.
A bug was fixed which for more than 1310 dofs
per block on the GPU caused an integer overflow.
In particular, the bug affected the pointer
calulation for the histograms. For 50 bins,
the offset of the current histogram is
50**2 * 1310**2 = 4,29025e9, which is just
lower than 2**32 for unsigned int.
50**2 * 1311**2 = 4,2968025e9, on the other
hand, is juast larger than 2**32, causing the
overflow. Modern GPUs with their large VRAM
(e.g. 24G) are well capable of reaching these
number of dofs per gpu block.

+++ src/PARENT_GPU/PARENT_GPU.cu
@@ -635,8 +635,8 @@ class Work{
             gpuErrchk( cudaSetDevice(gpu->id) );
             gpuErrchk( cudaMemset(gpu->layout->histograms, 0, bytes_to_zero) );
 
-        for (unsigned int i = 0; i < block1->n_dofs; i++) {
-          for (unsigned int j = 0; j < block2->n_dofs; j++) {
+        for (size_t i = 0; i < block1->n_dofs; i++) {
+          for (size_t j = 0; j < block2->n_dofs; j++) {
             
             unsigned int dof1_g = i + block1->dof_id_start_g;
             unsigned int dof2_g = j + block2->dof_id_start_g;
@@ -664,8 +664,8 @@ class Work{
         }
         
 
-        for (unsigned int i = 0; i < block1->n_dofs; i++) {
-          for (unsigned int j = 0; j < block2->n_dofs; j++) {
+        for (size_t i = 0; i < block1->n_dofs; i++) {
+          for (size_t j = 0; j < block2->n_dofs; j++) {
             unsigned int dof1_g = i + block1->dof_id_start_g;
             unsigned int dof2_g = j + block2->dof_id_start_g;
             PRECISION min1 = cpu_ram_layout->minima[dof1_g];
@@ -765,8 +765,8 @@ class Work{
             gpuErrchk( cudaSetDevice(gpu->id) );
             gpuErrchk( cudaMemset(gpu->layout->histograms, 0, bytes_to_zero) );
 
-        for (unsigned int i = 0; i < block->n_dofs - 1; i++) {
-          for (unsigned int j = i + 1; j < block->n_dofs; j++) {
+        for (size_t i = 0; i < block->n_dofs - 1; i++) {
+          for (size_t j = i + 1; j < block->n_dofs; j++) {
             unsigned int dof1_g = i + block->dof_id_start_g;
             unsigned int dof2_g = j + block->dof_id_start_g;
             PRECISION min1 = cpu_ram_layout->minima[dof1_g];
@@ -793,8 +793,8 @@ class Work{
         }
 
 
-        for (unsigned int i = 0; i < block->n_dofs - 1; i++) {
-          for (unsigned int j = i + 1; j < block->n_dofs; j++) {
+        for (size_t i = 0; i < block->n_dofs - 1; i++) {
+          for (size_t j = i + 1; j < block->n_dofs; j++) {
             unsigned int dof1_g = i + block->dof_id_start_g;
             unsigned int dof2_g = j + block->dof_id_start_g;
             PRECISION min1 = cpu_ram_layout->minima[dof1_g];
